@@ -2,17 +2,17 @@
   "name": "ultimate-tournament-client",
   "private": true,
   "dependencies": {
-    "babel": "^6.5.2",
-    "babel-cli": "^6.10.1",
-    "babel-core": "^6.10.4",
-    "babel-loader": "^6.2.4",
-    "babel-polyfill": "^6.9.1",
-    "babel-preset-es2015": "^6.9.0",
-    "babel-preset-react": "^6.9.0",
-    "babel-preset-stage-0": "^6.5.0",
-    "babel-runtime": "^6.9.2",
+    "babel-cli": "^6.24.1",
+    "babel-core": "^6.24.1",
+    "babel-loader": "^6.4.1",
+    "babel-polyfill": "^6.23.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-0": "^6.24.1",
+    "babel-runtime": "^6.23.0",
     "classnames": "^2.2.3",
     "d3": "3.5.17",
+    "date-arithmetic": "^3.1.0",
     "es5-shim": "^4.5.9",
     "events": "^1.1.1",
     "expose-loader": "^0.7.1",
@@ -23,20 +23,30 @@
     "leaflet-editable": "1.0.0-rc.2",
     "loader-utils": "^0.2.15",
     "lodash": "^4.13.1",
+    "moment": "^2.15.2",
+    "moment-range": "^2.2.0",
     "object-squish": "^1.0.0",
+    "prop-types": "^15.5.10",
     "query-string": "^3.0.0",
-    "react": "^15.2.1",
-    "react-bootstrap": "^0.30.2",
+    "react": "^15.4.1",
+    "react-bootstrap": "^0.30.8",
     "react-confirm": "^0.1.13",
-    "react-dom": "^15.2.1",
+    "react-datepicker": "^0.51.0",
+    "react-dnd": "^2.4.0",
+    "react-dnd-html5-backend": "^2.4.1",
+    "react-dom": "^15.4.1",
     "react-on-rails": "6.1.2",
+    "react-grid-layout": "^0.13.5",
+    "react-on-rails": "8.0.5",
     "react-overlays": "^0.6.0",
     "resolve-url-loader": "^1.4.3",
     "set-query-string": "^2.1.0",
-    "webpack": "^1.13.1"
+    "webpack": "^2.3.3",
+    "webpack-manifest-plugin": "^1.1.0",
+    "webpack-merge": "^4.1.0"
   },
   "devDependencies": {
-    "babel-eslint": "^7.0.0",
+    "babel-eslint": "^7.2.1",
     "babel-plugin-react-transform": "^2.0.2",
     "eslint": "^3.7.1",
     "eslint-config-standard": "^6.2.0",
@@ -45,11 +55,11 @@
     "eslint-plugin-standard": "^2.0.1",
     "estraverse-fb": "^1.3.1",
     "react-transform-hmr": "^1.0.4",
-    "webpack-dev-server": "^1.14.1"
+    "webpack-dev-server": "^2.4.2"
   },
   "scripts": {
     "build:production": "NODE_ENV=production webpack --config webpack.client.rails.build.config.js",
     "build:dev": "webpack --config webpack.client.rails.build.config.js",
-    "build:hot": "babel-node server-rails-hot.js"
+    "hot-assets": "NODE_ENV=development babel-node server-rails-hot.js"
   }
 }
