@@ -90,7 +90,7 @@ class Game < ApplicationRecord
   end
 
   def scheduled?
-    field_id && start_time
+    !!(field_id && start_time)
   end
 
   def played?
@@ -99,6 +99,7 @@ class Game < ApplicationRecord
   end
 
   def end_time
+    return unless start_time
     start_time + tournament.time_cap.minutes
   end
 
