@@ -1,8 +1,8 @@
-import IndexBase from './index_base'
-import FilterBar from './filter_bar'
-import filterFunction from '../modules/filter_function'
-import LinkCell from './link_cell'
-import FieldsStore from '../stores/fields_store'
+import IndexBase from './IndexBase'
+import FilterBar from './FilterBar'
+import filterFunction from '../modules/FilterFunction'
+import LinkCell from './LinkCell'
+import FieldsStore from '../stores/FieldsStore'
 
 class FieldsIndex extends IndexBase {
   constructor (props) {
