@@ -1,17 +1,25 @@
-/* eslint no-var: 0, no-console: 0 */
+/* eslint no-var: 0, no-console: 0, import/no-extraneous-dependencies: 0 */
 
-import webpack from 'webpack'
-import WebpackDevServer from 'webpack-dev-server'
+const webpack = require('webpack')
+const WebpackDevServer = require('webpack-dev-server')
+const webpackConfig = require('./webpack.client.rails.hot.config')
+const webpackConfigLoader = require('react-on-rails/webpackConfigLoader')
 
-import webpackConfig from './webpack.client.rails.hot.config'
-
-const hotRailsPort = process.env.HOT_RAILS_PORT || 3500
+const { resolve } = require('path')
+const configPath = resolve('..', 'config')
+const { hotReloadingUrl, hotReloadingPort, hotReloadingHostname } = webpackConfigLoader(configPath)
 
 const compiler = webpack(webpackConfig)
 
 const devServer = new WebpackDevServer(compiler, {
-  contentBase: `http://lvh.me:${hotRailsPort}`,
-  publicPath: webpackConfig.output.publicPath,
+  proxy: {
+    '*': `http://lvh.me:${hotReloadingPort}`
+  },
+  headers: {
+    'Access-Control-Allow-Origin': '*'
+  },
+  disableHostCheck: true,
+  contentBase: hotReloadingUrl,
   hot: true,
   inline: true,
   historyApiFallback: true,
@@ -27,9 +35,9 @@ const devServer = new WebpackDevServer(compiler, {
   }
 })
 
-devServer.listen(3500, 'localhost', err => {
+devServer.listen(hotReloadingPort, hotReloadingHostname, (err) => {
   if (err) console.error(err)
   console.log(
-    `=> 🔥  Webpack development server is running on port ${hotRailsPort}`
+    `=> 🔥  Webpack development server is running on ${hotReloadingUrl}`
   )
 })
