@@ -2,28 +2,15 @@
 // cd client && npm run build:client
 // Note that Foreman (Procfile.dev) has also been configured to take care of this.
 
+const merge = require('webpack-merge')
 const config = require('./webpack.client.base.config')
+const { resolve } = require('path')
+const webpackConfigLoader = require('react-on-rails/webpackConfigLoader')
+
+const configPath = resolve('..', 'config')
+const { webpackOutputPath, webpackPublicOutputDir } = webpackConfigLoader(configPath)
 
 const devBuild = process.env.NODE_ENV !== 'production'
-
-config.output = {
-  filename: '[name]-bundle.js',
-  path: '../app/assets/webpack'
-}
-
-// See webpack.common.config for adding modules common to both the webpack dev server and rails
-
-config.module.loaders.push(
-  {
-    test: /\.jsx?$/,
-    loader: 'babel-loader',
-    exclude: /node_modules/
-  },
-  {
-    test: require.resolve('react'),
-    loader: 'imports?shim=es5-shim/es5-shim&sham=es5-shim/es5-sham'
-  }
-)
 
 if (devBuild) {
   console.log('Webpack dev build for Rails') // eslint-disable-line no-console
@@ -32,4 +19,34 @@ if (devBuild) {
   console.log('Webpack production build for Rails') // eslint-disable-line no-console
 }
 
-module.exports = config
+module.exports = merge(config, {
+  output: {
+    filename: '[name]-[hash].js',
+
+    // Leading and trailing slashes ARE necessary.
+    publicPath: `/${webpackPublicOutputDir}/`,
+    path: webpackOutputPath
+  },
+
+  // See webpack.client.base.config for adding modules common to both webpack dev server and rails
+
+  module: {
+    rules: [
+      {
+        test: /\.jsx?$/,
+        use: 'babel-loader',
+        exclude: /node_modules/
+      },
+      {
+        test: require.resolve('react'),
+        use: {
+          loader: 'imports-loader',
+          options: {
+            shim: 'es5-shim/es5-shim',
+            sham: 'es5-shim/es5-sham'
+          }
+        }
+      }
+    ]
+  }
+})
