@@ -3,8 +3,7 @@
 # Version of your assets, change this if you want to expire all your assets.
 Rails.application.config.assets.version = '1.0'
 
-# Add folder with webpack generated assets to assets.paths
-Rails.application.config.assets.paths << Rails.root.join("app", "assets", "webpack")
+Rails.application.config.assets.paths << Rails.root.join('client', 'node_modules')
 
 # Precompile additional assets.
 # application.js, application.css, and all non-JS/CSS in app/assets folder are already added.
@@ -13,15 +12,11 @@ Rails.application.config.assets.precompile += [
   'brochure.js',
   'brochure/vendor.js',
   'admin.css',
+  'admin.js',
   'admin/vendor.js',
   'schedule_pdf.css',
   'app.css',
   'app.js',
   'app/vendor.js',
   'internal.js',
-]
-
-type = ENV["REACT_ON_RAILS_ENV"] == "HOT" ? "non_webpack" : "static"
-Rails.application.config.assets.precompile += [
-  "admin_#{type}.js",
 ]
